requirements.txt
discord.py==1.3.3
firebase-admin==3.0.0
python-dotenv==0.12.0
